# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Discord Bot Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/discord/callback:
    get:
      operationId: discordAuthCallback
      tags: [auth]
      summary: Discord OAuth2 callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirects to dashboard
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guilds:
    get:
      operationId: getGuilds
      tags: [guilds]
      summary: Get guilds the user manages
      responses:
        "200":
          description: List of guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guild"
        "401":
          description: Unauthorized
...
  /guilds/{guildId}/config:
    get:
      operationId: getGuildConfig
      tags: [config]
      summary: Get guild configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildConfig"
        "404":
          description: Not found
    put:
      operationId: updateGuildConfig
      tags: [config]
      summary: Update guild configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuildConfigBody"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildConfig"
  /guilds/{guildId}/embeds:
    get:
      operationId: getEmbeds
      tags: [embeds]
      summary: List embeds for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of embeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmbedDesign"
    post:
      operationId: createEmbed
      tags: [embeds]
      summary: Create a new embed design
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmbedBody"
      responses:
        "201":
          description: Created embed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedDesign"
  /guilds/{guildId}/embeds/{embedId}:
    get:
      operationId: getEmbed
      tags: [embeds]
      summary: Get a specific embed
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: embedId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Embed design
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedDesign"
    put:
      operationId: updateEmbed
      tags: [embeds]
      summary: Update an embed design
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: embedId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmbedBody"
      responses:
        "200":
          description: Updated embed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedDesign"
    delete:
      operationId: deleteEmbed
      tags: [embeds]
      summary: Delete an embed design
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: embedId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /guilds/{guildId}/embeds/{embedId}/send:
    post:
      operationId: sendEmbed
      tags: [embeds]
      summary: Send an embed to a channel via the bot
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: embedId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmbedBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DiscordUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        discriminator:
          type: string
        avatar:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
      required:
        - id
        - username
        - discriminator
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        botPresent:
          type: boolean
        memberCount:
          type: ["integer", "null"]
      required:
        - id
        - name
        - botPresent
...
    EmbedField:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        inline:
          type: boolean
      required:
        - name
        - value
    EmbedDesign:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        name:
          type: string
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        footerText:
          type: ["string", "null"]
        footerIconUrl:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        authorIconUrl:
          type: ["string", "null"]
        authorUrl:
          type: ["string", "null"]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/EmbedField"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - guildId
        - name
        - fields
        - createdAt
        - updatedAt
    CreateEmbedBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        footerText:
          type: ["string", "null"]
        footerIconUrl:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        authorIconUrl:
          type: ["string", "null"]
        authorUrl:
          type: ["string", "null"]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/EmbedField"
      required:
        - name
    SendEmbedBody:
      type: object
      properties:
        channelId:
          type: string
      required:
        - channelId
    WelcomeConfig:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedId:
          type: ["string", "null"]
      required:
        - enabled
    GoodbyeConfig:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        embedId:
          type: ["string", "null"]
      required:
        - enabled
    ModerationConfig:
      type: object
      properties:
        enabled:
          type: boolean
        logChannelId:
          type: ["string", "null"]
        muteRoleId:
          type: ["string", "null"]
      required:
        - enabled
    LoggingConfig:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        logJoins:
          type: boolean
        logLeaves:
          type: boolean
        logMessageDeletes:
          type: boolean
        logMessageEdits:
          type: boolean
        logModActions:
          type: boolean
      required:
        - enabled
        - logJoins
        - logLeaves
        - logMessageDeletes
        - logMessageEdits
        - logModActions
    AutoRoleConfig:
      type: object
      properties:
        enabled:
          type: boolean
        roleIds:
          type: array
          items:
            type: string
      required:
        - enabled
        - roleIds
    AntiSpamConfig:
      type: object
      properties:
        enabled:
          type: boolean
        maxMessages:
          type: integer
        timeWindowSeconds:
          type: integer
        action:
          type: string
      required:
        - enabled
        - maxMessages
        - timeWindowSeconds
        - action
    ReactionRoleItem:
      type: object
      properties:
        emoji:
          type: string
        roleId:
          type: string
      required:
        - emoji
        - roleId
    ReactionRolesConfig:
      type: object
      properties:
        enabled:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReactionRoleItem"
      required:
        - enabled
        - items
    GuildConfig:
      type: object
      properties:
        guildId:
          type: string
        prefix:
          type: string
        welcome:
          $ref: "#/components/schemas/WelcomeConfig"
        goodbye:
          $ref: "#/components/schemas/GoodbyeConfig"
        moderation:
          $ref: "#/components/schemas/ModerationConfig"
        logging:
          $ref: "#/components/schemas/LoggingConfig"
        autoRole:
          $ref: "#/components/schemas/AutoRoleConfig"
        antiSpam:
          $ref: "#/components/schemas/AntiSpamConfig"
        reactionRoles:
          $ref: "#/components/schemas/ReactionRolesConfig"
        updatedAt:
          type: string
      required:
        - guildId
        - prefix
        - welcome
        - goodbye
        - moderation
        - logging
        - autoRole
        - antiSpam
        - reactionRoles
    UpdateGuildConfigBody:
      type: object
      properties:
        prefix:
          type: string
        welcome:
          $ref: "#/components/schemas/WelcomeConfig"
        goodbye:
          $ref: "#/components/schemas/GoodbyeConfig"
        moderation:
          $ref: "#/components/schemas/ModerationConfig"
        logging:
          $ref: "#/components/schemas/LoggingConfig"
        autoRole:
          $ref: "#/components/schemas/AutoRoleConfig"
        antiSpam:
          $ref: "#/components/schemas/AntiSpamConfig"
        reactionRoles:
          $ref: "#/components/schemas/ReactionRolesConfig"
...